Revert "Move fade more local to avoid fading twice"
This reverts commit d0421fe20667530bf1bca1a5c8e3f387dde0cf6a
and fixes https://github.com/ppy/osu/issues/26801.

https://github.com/ppy/osu/pull/26703#discussion_r1469409667
was correct in saying that the early fade-out needs to be restored, and
that's because of the early-return. Legacy judgements that are the
temporary displayed judgement from new piece should also receive the
fade-out, and d0421fe20667530bf1bca1a5c8e3f387dde0cf6a broke that.