Revert incorrect end position optimisation
Closes https://github.com/ppy/osu/issues/26867.

Reverts fb92aaf0c032ec4615960d35fef2bb2e2894bfc6
and 2d0cfed30f87239ae74c5f0714fcb208ee8ed0db.

The applied optimisation may have been valid as long as it was
constrained to `Slider`. But it is not, as `SliderTailCircle` stores a
local copy of the object position. And as the commit message of
2d0cfed30f87239ae74c5f0714fcb208ee8ed0db states, this could be bypassed
by some pretty hacky delegation from `SliderTailCircle.Position` to the
slider, but it'd also be pretty hacky because it would make flows like
`PositionBindable` break down.

Long-term solution is to probably remove bindables from hitobjects.