Use better messaging for selected submission failure reasons
These have been cropping up rather often lately, mostly courtesy of
linux users, but not only:

	https://github.com/ppy/osu/issues/26840
	https://github.com/ppy/osu/issues/27008
	https://github.com/ppy/osu/discussions/26962

so this is a proposal for slightly improved messaging for such cases to
hopefully get users on the right track.

The original error is still logged to network log, so there's no
information loss.