Move flags log within if statement clause
In my previous commit I figured that I should put it outside of the if
statement so that it logs the compatibility mode flags regardless (in
case the other messagebox is shown for some reason and compatibility
mode is enabled). But I figured that it's highly likely this would never
happen, so might as well just keep it in the if statement.