Fix memory leaks resulting from loudness normalization
So I forgot to free the stream when everything's finished. This would cause the BASS stream to exist eternally, resulting in the garbage collector never collecting this, resulting in a ton of ram usage. Additionally, 60k bytes is a lot for something like this. So I reduced the size of the buffer to 10k bytes.