Fix bug with totalcount
So while I was looking through the diff, I noticed that I could simply
move the processedcount incrementor above the if...continue statement
and repurpose the notification to being "Verifying loudness levels"
instead of "Calculating loudness levels". This is an easy fix that still
offers a sense of transparency to the user.