Slightly refactor audio normalization background process
I've made it so that the beatmaps update automatically after their loudness information is calculated instead of deferring all the updates until everything has been calculated. I mainly did this to try and get the garbage collector to realize that the old values are no longer needed so they can be collected, but even after doing this the garbage collector seems to be stubborn and wants to hold onto the already processed objects.. I have no idea why. Regardless, I'm going to commit this since having them update immediately instead of being deferred is nice