Fix memory leaks
So I've found the culprit of the memory leak, bass. I made a typo when
writing this and did `SampleFree` instead of `StreamFree`. And due to
the lack of error handling here, I had no idea that `SampleFree` was
failing due to a mismatch in handle type.

This fixes that and also adds error handling to notify if freeing the
stream fails (but since freeing the stream isn't fatal, it doesn't do
anything, just logs)