Make PopulateSet void
So I noticed that reassigning beatmapset was kinda useless since we're
already editing it in place. I'm not really sure why the old PR did
this, and I mostly just copied this implementation from the old PR. I
assume that something changed since 2022 which allows us to not have to
reassign BeatmapSetInfo.

Regardless, stuff works without having to return a value, and I'm happy
with that.