Put sleep function call below null check
We skip if the current beatmap has audio normalization values. If not,
we process it. We don't really care about skipping since that has no
overhead, we care about processing since that has overhead. So moving
the sleep function under the null check means we can get a bit of
skipping in before pausing the processing.