Fix resume overlay not appearing after pausing inside window but outside of actual playfield area
Related to https://github.com/ppy/osu/discussions/27871 (although does
not actually fix the issue with the pause button, _if_ it is to be
considered an issue - the problem there is that the gameplay cursor
gets hidden, so the other condition in the modified check takes over).

Regressed in
https://github.com/ppy/osu/commit/bce3bd55e5a863e52f41598c306a248a79638843.
Reasoning for breakage is silent change in `this` when moving the
`Contains()` check (`DrawableRuleset` will encompass screen bounds,
while `OsuResumeOverlay` is only as big as the actual playfield).