Rework SkinnableDrawable to support single-instancing
Rather than defining a separate `DefaultStageConfiguration` and
`LegacyStageConfiguration`, we can create a single-instance
`SkinnableDrawable` which is reused.

The implementation of `reload()` is a little bit temporary at this
stage. The follow-up is to add a `Skin.ConfigureComponent()` or
`Skin.GetConfiguration()`.

Furthermore, the implementations of `Add()` and `Remove()` are no longer
(/were never?) required because the types in this "container" never
change. Undo/redo is handled directly by `SkinEditorChangeHandler`.