Replace base lookups with GetConfiguration()
The most contentious thing in the PR has been the introduction of the
`base.GetDrawableComponent()` calls. This follows in the footsteps of
"attempt 2" and adds a method specific to configuration.

One question I had to answer is "what about the defaults". Consider the
following:
- The classic skin is curently selected which positions ManiaPlayfield
  with TopLeft anchors and somewhere towards the left of the screen.
- The argon skin is selected which wants Centre anchors and a zeroed
  position.
- A third hypothetical, is that the triangles skin is selected that wants
  totally separate defaults from all of those.

Attempt 2 handled this by exposing an `ApplyDefaults()` method on the
`IConfigurableDrawable` (in this case, `ManiaPlayfield` itself). This
method was implemented something as such:

```
Anchor = Anchors.TopLeft;
Origin = Anchors.TopLeft;
Position = Vector2.Zero;
Rotation = 0;
```

Now I ask you, dear reader, to tell me why these specific properties
need to be reset. Why should `Scale` not also? What about any of the
many other properties of `Drawable`? This logic presents an issue to
me - that there is something, _somewhere_, doing magical things and
we somehow need to be omniscient of that.

In this case, we need to know what each skin transformer wants, as well
as what what the `SkinEditor` is capable of doing. It turns out `Scale`
not being reset is a mistake. Whoops!

I foresee this going wrong in the future as well. If the `SkinEditor`
receives some additional functionality, we will conscious of the fact
that _every single implementation_ will need to be updated to reset that
property. Every usage is _required_ to have this implementation.

A second minor issue, to me, is that this looks very awkward. I imagine
the question would come up at some point of why we need this when we have
the constructor in the first place.
I think that's a very good question, for which the only answer I have is
that "it be like it do". One good property about my original
implementation in this PR is that the `Drawable` was always
reconstructed and had control of its own defaults.
This commit is still not quite as "neat" as that, in my opinion, but at
least it avoids the bigger issue above.