Improve error handling, nullability, and optionality across the board
So I decided to improve the error handling since a lot of stuff is
confusing, and I'd like for it to be more verbose.

So this commit changes the IntegratedLoudness values in both
`AudioNormalization` and `BassAudioNormalization` to be optional. I did
this because the loudness calculation values can fail to be calculated,
and having to check this by checking for the value of "1" is
disorganized. I decided to change it so that null = couldnt calculate
because of an error. I also made this change because the code would only
check if the class exists on the BeatmapInfo, not that the
IntegratedLoudness actually exists. In my previous code, which would
return if there was an error before setting the IntegratedLoudness, this
would mean that the class exists on the BeatmapInfo, but the
IntegratedLoudness is still null/uninitialized.

I hope that this approach helps make it obvious when IntegratedLoudness
might be null due to an error, and also helps make it obvious when it
does have a value.