Fix improper cache invalidation
So I originally made this invalidate the current BeatmapInfo because I
forgot that we would later go on to populate other difficulties. This
resulted in certain difficulties not having their caches invalidated.
The way this loop works is find a difficulty for a beatmapset ->
calculate loudness of that beatmap -> go through the set and find
difficulties with the same audio file then copy the loudness information
to them -> go through the rest of the difficulties that dont have a
loudness value -> go to next set
This means that difficulties which had their values populated wouldn't
have their caches invalidated since only the difficulty which was
initially calculated would be invalidated. By doing this, we effectively
invalidate the set's cache every time we update the loudness information
for it.